# backend/openapi.yaml
openapi: 3.1.0
info:
  title: BudgetBuddy API
  version: "1"
  description: |
    Operational header behavior: responses include `X-Request-Id`.
    If request provides `X-Request-Id`, the API echoes it; otherwise it generates one.
servers:
- url: /api
security:
- BearerAuth: []
paths:
...
  /categories:
    get:
      summary: List categories
      description: |
        Cursor pagination uses an opaque `base64url(JSON)` token.
        Ordering is deterministic by `created_at DESC, id DESC`, and `next_cursor` is derived from that same sort key.
        For stable datasets, paging is deterministic with no overlaps or skips; terminal page returns `next_cursor: null`.
        Under concurrent mutations, pagination is best-effort and does not guarantee snapshot semantics.
      parameters:
      - name: include_archived
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: type
        in: query
        required: false
        schema:
          type: string
          enum: [ income, expense ]
      - name: cursor
        in: query
        required: false
        description: Opaque `base64url(JSON)` token built from `created_at` and `id` of the last item in the previous page.
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                type: object
                required: [ items, next_cursor ]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
                  next_cursor:
                    type: [ string, "null" ]
        "400":
          description: Invalid cursor
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

    post:
      summary: Create category
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
      responses:
        "201":
          description: Created
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Category name already exists for this type
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /categories/{category_id}:
    parameters:
    - name: category_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

    get:
      summary: Get category
      responses:
        "200":
          description: OK
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

    patch:
      summary: Update category
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Category name already exists for this type
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

    delete:
      summary: Archive category
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  headers:
    X-Request-Id:
      description: Correlation id for tracing requests. Propagated when provided by client, generated otherwise.
      schema:
        type: string
...
  schemas:
    ProblemDetails:
      type: object
      required: [ type, title, status ]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
...
    Category:
      type: object
      required: [ id, name, type, archived_at ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [ income, expense ]
        note:
          type: string
        archived_at:
          type: [ string, "null" ]
          format: date-time
    CategoryCreate:
      type: object
      required: [ name, type ]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [ income, expense ]
        note:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        note:
          type: string
        archived_at:
          type: [ string, "null" ]
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT